+++ server/subscriptions/sql/index.ts
@@ -16,12 +16,30 @@ export const getTriggeredThreadSubscriptions = `
 export const getSubscriptionsWebhooks = `
     SELECT 
         webhook,
+        handler_type AS webhook_handler_type,
         array_agg(DISTINCT subscriptions.string_id) AS subscription_ids
     FROM subscriptions
     LEFT JOIN subscription_webhooks sw ON subscriptions.id = sw.subscription_id
     LEFT JOIN webhooks ON sw.webhook_id = webhooks.id
     WHERE subscriptions.string_id = ANY($/subscriptions_string_ids/)
-    GROUP BY webhook
+    GROUP BY webhook, webhook_handler_type
+`;
+
+const getTriggeredWebhooks = `
+    SELECT
+        webhook,
+        handler_type AS webhook_handler_type,
+        array_agg(DISTINCT subscriptions.string_id) AS subscription_ids,
+        array_agg(DISTINCT subscriptions.name) AS subscription_names,
+        array_agg(DISTINCT categories.category) AS triggered_categories
+    FROM subscriptions
+        LEFT JOIN board_category_subscriptions bcs ON bcs.subscription_id = subscriptions.id
+        LEFT JOIN boards ON bcs.board_id = boards.id
+        LEFT JOIN categories ON bcs.category_id = categories.id
+        LEFT JOIN subscription_webhooks sw ON subscriptions.id = sw.subscription_id
+        LEFT JOIN webhooks ON sw.webhook_id = webhooks.id
+    WHERE boards.slug = $/board_slug/ AND categories.category = ANY($/category_names/)
+    GROUP BY webhook, webhook_handler_type
 `;
 
 export default {
@@ -30,4 +48,5 @@ export default {
   ),
   getTriggeredThreadSubscriptions,
   getSubscriptionsWebhooks,
+  getTriggeredWebhooks,
 };
